captioning-service/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
redis==5.0.1
pydantic==2.5.0
httpx==0.25.2

# JoyCaption / LLaVA dependencies
# PyTorch 2.4+ required by transformers>=4.40.0
torch>=2.4.0
transformers>=4.40.0
accelerate>=0.26.0
bitsandbytes>=0.42.0
sentencepiece>=0.1.99
protobuf>=3.20.3
pillow>=10.1.0
# Pin numpy<2.0 to avoid compatibility issues with torch
numpy>=1.24.3,<2.0
opencv-python>=4.8.1.78
